--- GetYourWay/.settings/org.eclipse.wst.jsdt.ui.superType.container
@@ -1,1 +1,0 @@
-org.eclipse.wst.jsdt.launching.baseBrowserLibrary
