+++ GetYourWay/src/main/java/com/qa/security/SecController.java
@@ -1,0 +1,5 @@
+package com.qa.security;
+
+public class SecController {
+
+}
