+++ GetYourWay/src/main/resources/log4j.xml
