+++ GetYourWay/src/main/webapp/TestSearch.html
@@ -1,0 +1,67 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta name="viewport" content="initial-scale=1.0, user-scalable=no">
+    <meta charset="utf-8">
+    <title>Directions service</title>
+    <style>
+      html, body, #map-canvas {
+        height: 100%;
+        margin: 0px;
+        padding: 0px
+      }
+      #panel {
+        position: absolute;
+        top: 5px;
+        left: 50%;
+        margin-left: -180px;
+        z-index: 5;
+        background-color: #fff;
+        padding: 5px;
+        border: 1px solid #999;
+      }
+    </style>
+<script type="text/javascript"
+      src="https://maps.googleapis.com/maps/api/js?key=">
+<script src="C:/getyourway/GetYourWayTemp/src/main/webapp/javaScripts/googleStuff.js"></script>
+
+</head>
+<body>
+
+<div id="panel">
+<strong>Start: </strong>
+<select id="start" onchange="calcRoute();">
+  <option value="chicago, il">Chicago</option>
+  <option value="st louis, mo">St Louis</option>
+  <option value="joplin, mo">Joplin, MO</option>
+  <option value="oklahoma city, ok">Oklahoma City</option>
+  <option value="amarillo, tx">Amarillo</option>
+  <option value="gallup, nm">Gallup, NM</option>
+  <option value="flagstaff, az">Flagstaff, AZ</option>
+  <option value="winona, az">Winona</option>
+  <option value="kingman, az">Kingman</option>
+  <option value="barstow, ca">Barstow</option>
+  <option value="san bernardino, ca">San Bernardino</option>
+  <option value="los angeles, ca">Los Angeles</option>
+</select>
+<strong>End: </strong>
+<select id="end" onchange="calcRoute();">
+  <option value="chicago, il">Chicago</option>
+  <option value="st louis, mo">St Louis</option>
+  <option value="joplin, mo">Joplin, MO</option>
+  <option value="oklahoma city, ok">Oklahoma City</option>
+  <option value="amarillo, tx">Amarillo</option>
+  <option value="gallup, nm">Gallup, NM</option>
+  <option value="flagstaff, az">Flagstaff, AZ</option>
+  <option value="winona, az">Winona</option>
+  <option value="kingman, az">Kingman</option>
+  <option value="barstow, ca">Barstow</option>
+  <option value="san bernardino, ca">San Bernardino</option>
+  <option value="los angeles, ca">Los Angeles</option>
+</select>
+</div>
+
+
+
+</body>
+</html>
