+++ GetYourWay/src/main/webapp/WEB-INF/views/FrontPage.html
@@ -1,0 +1,25 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
+<HTML>
+<HEAD>
+<TITLE>Index Page</TITLE>
+</HEAD>
+<BODY>
+
+<H1></H1>
+
+<DIV>
+
+</DIV>
+
+<DIV>
+<UL>
+<LI><A href="home">Home</A></LI>
+<LI><A href="login">Login</A></LI>
+<LI><A href="aboutus">About Us</A></LI>
+<LI><A href="feedback">Feedback</A></LI>
+</UL>
+</DIV>
+
+
+</BODY>
+</HTML>
