+++ GetYourWay/src/main/webapp/WEB-INF/views/register.jsp
@@ -1,0 +1,34 @@
+<DIV align="left">
+<H2>Please type in your details:</H2>
+
+<FORM method="get" action="registerdetails">
+<TABLE border="0">
+<TR>
+<TD width="100" align="right"><B>Name: </B></TD>
+<TD><INPUT type="text" name="fullname"></TD>
+</TR>
+<TR>
+<TD width="100" align="right"><B>Username: </B></TD>
+<TD><INPUT type="text" name="username"></TD>
+</TR>
+<TR>
+<TD width="100" align="right"><B>Age: </B></TD>
+<TD><INPUT type="text" name="age"></TD>
+</TR>
+<TR>
+<TD width="100" align="right"><B>Email: </B></TD>
+<TD><INPUT type="text" name="email"></TD>
+</TR>
+</TABLE>
+<BR><BR>
+</DIV>
+<DIV align="left">
+<H2>Would you like to enter a default departure location?</H2>
+
+</DIV>
+<DIV align="left">
+<INPUT type="submit" value="REGISTER"></TD><TD></TD>
+</FORM>
+</DIV>
+
+<BR><BR><BR>
