--- GetYourWayTemp/src/main/java/com/qa/security/SecController.java
@@ -1,5 +1,0 @@
-package com.qa.security;
-
-public class SecController {
-
-}
