--- GetYourWayTemp/src/main/webapp/SearchPage.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-<title>Front Page</title>
-</head>
-<body>
-<h1>Enter your Search</h1>
-<div>
-Please enter the day ranges of the Festival that interest you:
-<form action="googleMaps.jsp" method="POST">
-<br>
-Origin:
-<input type="TEXT" size="10" name="origin">
-<br>
-Destination:
-<input type="TEXT" name="destination">
-<input type="SUBMIT" value="Submit">
-</form>
-</div>
-</body>
-</html>
