--- GetYourWayTemp/src/main/webapp/WEB-INF/views/Login.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="ISO-8859-1">
-<title>Insert title here</title>
-</head>
-<body>
-
-<form action='j_spring_security_check' method="POST">
-<table>
-	<tr>
-		<td>Username:</td>
-		<td><input type='text' name='j_username'></td>
-	</tr>
-	<tr>
-		<td>Password:</td>
-		<td><input type='password' name='j_password'></td>
-	</tr>
-	<tr>
-		<td><input name="reset" type="reset" value="Clear"></td>
-		<td><input name="submit" type="submit" value="Login"></td>
-	</tr>
-</table>
-</form>
-
-</body>
-</html>
