--- GetYourWayTemp/src/main/webapp/WEB-INF/views/footer.jsp
@@ -1,11 +1,0 @@
-<BR><BR><HR><BR>
-
-<DIV>
-Contact details: 
-<A href="mailto:<%= application.getInitParameter("email") %>"><%= application.getInitParameter("email") %></A>.
-<BR>This page was processed at <%= new java.util.Date() %>.
-</DIV>
-
-</body>
-
-</html>
